streamlit/requirements.txt
numpy==1.24.3
streamlit==1.22.0
Pillow==9.5.0
tensorflow==2.19.0
pyarrow==11.0.0
opencv-python-headless
keras==3.12.0

